Labfiles/08-gen-ai-vision/python/requirements.txt
dotenv 
azure-identity
azure-ai-projects==1.0.0b9
azure-ai-inference==1.0.0b9
openai